requirements.txt
flask
opencv-python
numpy
kafka-python
pyspark[sql,streaming,kafka]
face-recognition
cassandra-driver
libevdev
cffi